third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGMarkerElement-dom-markerWidth-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'markerWidth' attribute of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.getAttribute('markerWidth') is "0"
PASS markerElement.getAttribute('markerWidth') is "2"
PASS successfullyParsed is true

TEST COMPLETE
